feat(softwareRank): softwarerank move to shared-components
softwarerank move to shared-components
