+++ packages/shared-components/src/assets/pic/gold.png
