+++ packages/shared-components/src/assets/pic/sliver.png
