+++ packages/shared-components/src/assets/pic/white.png
